python/translateText/requirements.txt
six
google-cloud-translate
